Experiments

+++ src/llatvfl/core/model.h
@@ -9,6 +9,6 @@ template <typename PartyName>
 struct TreeModelBase
 {
     TreeModelBase(){};
-    virtual void fit(vector<PartyName> &parties, vector<double> &y) = 0;
-    virtual vector<double> predict_raw(vector<vector<double>> &X) = 0;
+    virtual void fit(vector<PartyName> &parties, vector<float> &y) = 0;
+    virtual vector<float> predict_raw(vector<vector<float>> &X) = 0;
 };
