+++ src/llatvfl/core/party.h
@@ -19,24 +19,24 @@ using namespace std;
 
 struct Party
 {
-    vector<vector<double>> x; // a feature vector of this party
-    vector<int> feature_id;   // id of the features
-    int party_id;             // id of this party
+    vector<vector<float>> x; // a feature vector of this party
+    vector<int> feature_id;  // id of the features
+    int party_id;            // id of this party
     int min_leaf;
-    double subsample_cols; // ratio of subsampled columuns
+    float subsample_cols; // ratio of subsampled columuns
     bool use_missing_value;
     int seed;
 
     int col_count; // the number of columns
     int subsample_col_count;
 
-    unordered_map<int, tuple<int, double, int>> lookup_table; // record_id: (feature_id, threshold, missing_value_dir)
+    unordered_map<int, tuple<int, float, int>> lookup_table; // record_id: (feature_id, threshold, missing_value_dir)
     vector<int> temp_column_subsample;
-    vector<vector<double>> temp_thresholds; // feature_id->threshold
+    vector<vector<float>> temp_thresholds; // feature_id->threshold
 
     Party() {}
-    Party(vector<vector<double>> &x_, vector<int> &feature_id_, int &party_id_,
-          int min_leaf_, double subsample_cols_,
+    Party(vector<vector<float>> &x_, vector<int> &feature_id_, int &party_id_,
+          int min_leaf_, float subsample_cols_,
           bool use_missing_value_ = false, int seed_ = 0)
     {
         validate_arguments(x_, feature_id_, party_id_, min_leaf_, subsample_cols_);
@@ -49,11 +49,11 @@ struct Party
         seed = seed_;
 
         col_count = x.at(0).size();
-        subsample_col_count = int(subsample_cols * double(col_count));
+        subsample_col_count = int(subsample_cols * float(col_count));
     }
 
-    void validate_arguments(vector<vector<double>> &x_, vector<int> &feature_id_, int &party_id_,
-                            int min_leaf_, double subsample_cols_)
+    void validate_arguments(vector<vector<float>> &x_, vector<int> &feature_id_, int &party_id_,
+                            int min_leaf_, float subsample_cols_)
     {
         try
         {
@@ -92,24 +92,24 @@ struct Party
         }
     }
 
-    unordered_map<int, tuple<int, double, int>> get_lookup_table()
+    unordered_map<int, tuple<int, float, int>> get_lookup_table()
     {
         return lookup_table;
     }
 
-    vector<double> get_threshold_candidates(vector<double> &x_col)
+    vector<float> get_threshold_candidates(vector<float> &x_col)
     {
-        vector<double> x_col_wo_duplicates = remove_duplicates<double>(x_col);
-        vector<double> thresholds(x_col_wo_duplicates.size());
+        vector<float> x_col_wo_duplicates = remove_duplicates<float>(x_col);
+        vector<float> thresholds(x_col_wo_duplicates.size());
         copy(x_col_wo_duplicates.begin(), x_col_wo_duplicates.end(), thresholds.begin());
         sort(thresholds.begin(), thresholds.end());
         return thresholds;
     }
 
-    bool is_left(int record_id, vector<double> &xi)
+    bool is_left(int record_id, vector<float> &xi)
     {
         bool flag;
-        double x_criterion = xi[feature_id[get<0>(lookup_table[record_id])]];
+        float x_criterion = xi[feature_id[get<0>(lookup_table[record_id])]];
         if (isnan(x_criterion))
         {
             try
@@ -158,12 +158,12 @@ struct Party
             missing_dir = 0;
         }
         int row_count = idxs.size();
-        vector<double> x_col(row_count);
+        vector<float> x_col(row_count);
         for (int r = 0; r < row_count; r++)
             x_col[r] = x[idxs[r]][feature_opt_id];
 
         vector<int> left_idxs;
-        double threshold = temp_thresholds[feature_opt_pos][threshold_opt_pos];
+        float threshold = temp_thresholds[feature_opt_pos][threshold_opt_pos];
         for (int r = 0; r < row_count; r++)
             if (((!isnan(x_col[r])) && (x_col[r] <= threshold)) ||
                 ((isnan(x_col[r])) && (missing_dir == 1)))
@@ -175,7 +175,7 @@ struct Party
     int insert_lookup_table(int feature_opt_pos, int threshold_opt_pos)
     {
         int feature_opt_id, missing_dir;
-        double threshold_opt;
+        float threshold_opt;
         feature_opt_id = temp_column_subsample[feature_opt_pos % subsample_col_count];
         threshold_opt = temp_thresholds[feature_opt_pos][threshold_opt_pos];
