+++ test/test_metric.cpp
@@ -9,10 +9,10 @@ using namespace std;
 int main()
 {
     vector<int> y_true = {0, 0, 0, 0, 1, 1, 1, 1};
-    vector<double> y_pred = {0.2, 0.3, 0.6, 0.8, 0.4, 0.5, 0.7, 0.9};
+    vector<float> y_pred = {0.2, 0.3, 0.6, 0.8, 0.4, 0.5, 0.7, 0.9};
 
     /*
-    pair<double, double> fpr_and_tpr = get_fpr_and_tpr(y_pred, y_true, 0.2);
+    pair<float, float> fpr_and_tpr = get_fpr_and_tpr(y_pred, y_true, 0.2);
     assert(fpr_and_tpr.first == 1 && fpr_and_tpr.second == 1);
     fpr_and_tpr = get_fpr_and_tpr(y_pred, y_true, 0.4);
     assert(fpr_and_tpr.first == 0.5 && fpr_and_tpr.second == 1);
