+++ test/test_utils.cpp
@@ -5,9 +5,9 @@
 
 int main()
 {
-    vector<double> in = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11};
-    vector<double> quartiles = Quantile<double>(in, {0.25, 0.5, 0.75});
-    vector<double> test_quartiles = {3.25, 6, 8.75};
+    vector<float> in = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11};
+    vector<float> quartiles = Quantile<float>(in, {0.25, 0.5, 0.75});
+    vector<float> test_quartiles = {3.25, 6, 8.75};
     for (int i = 0; i < quartiles.size(); i++)
     {
         assert(quartiles[i] == test_quartiles[i]);
